+++ src/main/java/com/mujeeb/barter/repository/CityRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.mujeeb.barter.entity.City;
 
-public interface CityRepository extends CrudRepository<City, Long> {
+public interface CityRepository extends CrudRepository<City, Integer> {
 
-	City findById(int id);
+	City findById(Long id);
 	
 	List<City> findAllCities();
